refactor: rebase extends methods for view from trait to abstract class
