+++ src/Views/IndexView.php
@@ -4,20 +4,9 @@ declare(strict_types=1);
 
 namespace Source\Views;
 
-use Source\App\Lib\Traits\TPageBuilder;
-
-class IndexView
+class IndexView extends View
 {
-    use TPageBuilder;
-
-    private string $header;
-
-    private string $headerPath;
-
-    private string $templatePath;
-
-
-    public function __construct(private array $domainConfig, private array $posts)
+    public function __construct(protected array $domainConfig, protected array $posts)
     {
         $this->headerPath = dirname(__DIR__) . '/Layouts/header.phtml';
         $this->templatePath = dirname(__DIR__) . '/Layouts/index.phtml';
